requirements.txt
flask
ultralytics
pillow
numpy
requests
python-dotenv
